+++ src/components/ui/Loader2.tsx
@@ -1,16 +1,19 @@
 import { Dots } from './Dots';
+import styles from './Loading.module.scss';
 
 // temp
 function Loader2() {
   return (
-    <span
-      style={{
-        display: 'flex',
-        margin: '15px 100%',
-      }}
-    >
-      Loading <Dots />
-    </span>
+    <div className={styles.container}>
+      <span
+        style={{
+          display: 'flex',
+          margin: '15px 100%',
+        }}
+      >
+        Loading <Dots />
+      </span>
+    </div>
   );
 }
