+++ src/components/ui/loading.tsx
@@ -1,16 +1,15 @@
-// eslint-disable-next-line import/prefer-default-export
-export function Loading() {
+import styles from './Loading.module.scss';
+
+function Loading() {
   return (
-    <div className="lds-grid">
-      <div />
-      <div />
-      <div />
-      <div />
-      <div />
-      <div />
-      <div />
-      <div />
-      <div />
+    <div className={styles.container}>
+      <div className={styles['lds-grid']}>
+        {Array.from(Array(9)).map((_, index) => (
+          <div key={index} />
+        ))}
+      </div>
     </div>
   );
 }
+
+export default Loading;
