+++ src/style/main.css
@@ -6,7 +6,6 @@
 @import './got.css';
 @import './home.css';
 
-@import './loading.css';
 @import './notFound.css';
 @import './popups.css';
